Refactor the code, add cache for tld metric values
