+++ edu.rice.cs.hpc.data/src/edu/rice/cs/hpc/data/db/version3/MetricValueCollection3.java
@@ -1,10 +1,9 @@
-package edu.rice.cs.hpc.data.experiment.metric.version3;
+package edu.rice.cs.hpc.data.db.version3;
 
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 
-import edu.rice.cs.hpc.data.db.DataSummary;
 import edu.rice.cs.hpc.data.experiment.BaseExperimentWithMetrics;
 import edu.rice.cs.hpc.data.experiment.Experiment;
 import edu.rice.cs.hpc.data.experiment.metric.BaseMetric;
