+++ edu.rice.cs.hpcdata.tld/src/edu/rice/cs/hpcdata/tld/plot/DataPlot.java
@@ -9,8 +9,7 @@ import java.nio.channels.FileChannel.MapMode;
 import java.util.ArrayList;
 import java.util.List;
 
-
-import edu.rice.cs.hpc.data.db.DataCommon;
+import edu.rice.cs.hpc.data.db.version3.DataCommon;
 
 /*******************************************************************************************
  *
